Create longest consecutive integers subsequence.py
Input:[ 3, 101, 1,  7, 6,  4,  8, 2, 102]
Output: 4
The explanation for this problem: The longest subsequence for consecutive integer numbers is [1, 2, 3, 4].

 There is more than one consecutive integer’s subsequence that we can extract from this array.
[101, 102] with 2 length
[6, 7, 8] with 3 length
[1, 2, 3, 4] with 4 length
And as we choose the longest consecutive subsequence, then the answer is : [ 1, 2, 3, 4] with 4  elements in length.
In this question, I try to write an algorithm to calculate the longest consecutive subsequence in the given array. First, I need to mention that the consecutive numbers are the numbers that are different from each other's by one, which means; we add "1" to get the next number in the sequence.

--------------------------------
Brute-Force Algorithm :
#This brute force algorithm to solve the problem to find the length of
#the longest consecutive integers subsequence
function findLongestConsecutiveSubsequence(arr):
    # put the initial value for longestArrayLength to 0
    longestArrayLength = 0
    # make Iteration among each item in arr
    for the item in arr:
        # initialize  the  currentIndexLength to 1
        currentIndexLength = 1
        # While item + 1 is in the arr, increment  the currentIndexLength and item
        while item + 1 in arr:
            currentIndexLength += 1
            item += 1
        # Ipgrade longestArrayLength if currentIndexLength is come to be greater than #longestArrayLength
        longestArrayLength = max(longestArrayLength, currentIndexLength)
    # Return the value of  longestArrayLength
    return longestArrayLength
   #call function in the output
findLongestConsecutiveSubsequence(arr)
---------------------------------
Analyzing the algorithm:
The worst case: this algorithm creates two nested loops to find the length of the longest subsequence array. Time complexity  O(n^2)
The average case: will be as same as the worst case
The best case: will be if the given array is sorted in ascending order, the time complexity is O(n)

-----------------------------------